Calculate the vector value at a given position

Calculate the vector value at a given position

Usage

# S3 method for class 'vectorfield'
predict(object, pos, linear_interp = FALSE, calculate_a = TRUE, ...)

Arguments

object

A vectorfield project generated by fit_2d_vf().

pos

A vector, the position of the vector.

linear_interp

Use linear interpolation method to estimate the drift vector (and the diffusion matrix). This can speed up the calculation. If TRUE, be sure that a linear grid was calculated for the vector field using <vf> <- add_interp_grid(<vf>).

calculate_a

Effective when linear_interp == TRUE. Do you want to calculate the diffusion matrix? Use FALSE can save some time.

...

Not in use.

Value

A list of v, the drift part that is used for vector fields, and a (when calculate_a == TRUE), the diffusion part at a given position.

See also